upgrade to wear os 3

+++ app/src/main/java/de/wagner_wedtlenstedt/evccwearosclient/data/EvccLoadpointModel.kt
@@ -1,7 +1,7 @@
 package de.wagner_wedtlenstedt.evccwearosclient.data
 
 data class EvccLoadpointModel(
-    var chargePower:Int?=0,
+    var chargePower:Float?=0.0f,
     var charging:Boolean?=false,
     var connected:Boolean?=false,
     var enabled:Boolean?=false,
