+++ app/src/main/java/de/wagner_wedtlenstedt/evccwearosclient/data/EvccResultModel.kt
@@ -2,14 +2,14 @@ package de.wagner_wedtlenstedt.evccwearosclient.data
 
 data class EvccResultModel(
     var batteryConfigured:Boolean?=false,
-    var batteryPower:Int?=0,
+    var batteryPower:Float?=0.0f,
     var batterySoC:Int?=0,
     var gridConfigured:Boolean?=false,
-    var gridPower:Int?=0,
-    var homePower:Int?=0,
+    var gridPower:Float?=0.0f,
+    var homePower:Float?=0.0f,
     var loadpoints:List<EvccLoadpointModel>?=null,
     var prioritySoC:Int?=0,
     var pvConfigured:Boolean?=false,
-    var pvPower:Int?=0,
+    var pvPower:Float?=0.0f,
     var siteTitle:String?=""
 )
